Virtualization_and_Cloud_Computing_Final_Assignment
Virtualization - Module Assignment

The purpose of this assignment is to give you a thorough understanding
of virtualisation and the implementation of VMware vSphere and common
tasks by deepening your practical knowledge and skills in relation to
both.
You will be asked to carry out a series of practical tasks.
Important: an essential part of the assignment will be to document the
procedures by which you completed each task from beginning to end.
Each task should be described in a series of numbered steps. In
addition, for each task, you should provide two or three legible screen
shots as evidence that you have actually worked through the task.
(Screen shots are normally placed at the most important stages/steps in
any task). The documentation should be laid out like a professional
user’s manual, aimed at the novice technician (someone who knows little
of such procedures). The manual should adopt the standards we associate
with such documentation.
Introduction
At the end of the assignment you will be required to produce the
following deliverables:
 A user’s manual documenting the steps you used to complete each tasks
outlined below. The documentation should be provided in both software
and hard copy. For each of the tasks, all technical issues with regard
to hardware and software requirements which must be met should be
detailed.
In this assignment you will be asked to create virtual machines.  On
each of these machines you will be asked to carry out a number of
practical tasks.